requirements.txt
pyaudio
deepgram-sdk
openai
